refactor: reorganize text model implementation and enhance logging
